+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [HEAD] Unreleased
+- prevent extra evaluations in bulk expression solver
+
 ## [v3.4.0] 2020-12-07
 - allow access to intermediate values of flattened hashes
 - catch invalid array syntax in the parse phase
@@ -12,6 +15,7 @@
 - add enum functions `ANY`, `ALL`, `MAP` and `PLUCK`
 - allow self-referential formulas in bulk expression solver
 - misc internal fixes and enhancements
+
 ## [v3.3.4] 2019-11-21
 - bugfix release
 
@@ -199,6 +203,7 @@
 ## [v0.1.0] 2012-01-20
 - initial release
 
+[HEAD]:  https://github.com/rubysolo/dentaku/compare/v3.4.0...HEAD
 [v3.4.0]:  https://github.com/rubysolo/dentaku/compare/v3.3.4...v3.4.0
 [v3.3.4]:  https://github.com/rubysolo/dentaku/compare/v3.3.3...v3.3.4
 [v3.3.3]:  https://github.com/rubysolo/dentaku/compare/v3.3.2...v3.3.3